move to_ethereum_proof/inputs to mopro-ffi
